feat(home): Separate hooks
