+++ src/features/home/Container.tsx
@@ -1,39 +1,11 @@
-import { FunctionComponent, useEffect, useRef, useState } from 'react';
+import { FunctionComponent } from 'react';
 
 import ContentContainer from './components/Container';
 import { data } from './constants';
+import useHomeContainer from './hooks/useHomeContainer';
 
 const HomeContainer: FunctionComponent = () => {
-  const refList = useRef<(HTMLDivElement | null)[]>([]);
-  const [activeIndex, setActiveIndex] = useState(0);
-
-  useEffect(() => {
-    if (!refList.current) return;
-
-    const io = new IntersectionObserver(
-      (entries, observer) => {
-        entries.forEach((entry, index) => {
-          if (entry.isIntersecting) {
-            const target = entry.target as HTMLDivElement;
-            const parsedId = Number(target.dataset.id);
-            if (isNaN(parsedId)) return;
-            setActiveIndex(parsedId);
-          }
-        });
-      },
-      { rootMargin: '-50% 0px' },
-    );
-    refList.current.forEach((ref) => {
-      if (!ref) return;
-      io.observe(ref);
-    });
-    return () => {
-      refList.current.forEach((ref) => {
-        if (!ref) return;
-        io.unobserve(ref);
-      });
-    };
-  }, [refList]);
+  const { refList, activeIndex } = useHomeContainer();
 
   return (
     <div className="home-container scroll-snap-container">
